moved script to make it easier to find
